Add a new struct wpe_view_backend_exportable_fdo_buffer
This is an opaque structure that encapsulates information about a
buffer delivered to the application for rendering.

Since the introduction of linux-dmabuf interface, there are now two
types of buffer that can be delivered: wayland buffer resource, and
linux-dmabuf buffer descriptors.

This structure, together with the single 'export_buffer' interface
callback that will be introduced in a later patch, abstracts the
underlying nature of the buffer to applications.