Add new public method wpe_view_backend_exportable_fdo_get_egl_image()
This method import the given opaque buffer into an EGLImageKHR.
It will choose the proper EGL attributes to use, based on the type of
buffer (wl_resource or linux-dmabuf).

Returns an EGLImageKHR, which must be destroyed by the application
by calling eglDestroyImageKHR(); or NULL on error, in which case
the application can use eglGetError() to retrieve it.