Add a new public wpe_view_backend_exportable_fdo_get_dmabuf_attributes()
This method retrieves the linux-dmabuf attributes of a given opaque
buffer.

This is a low-level function for applications that want the flexibility
to choose the API to import the buffer, and/or manipulate the attributes
before importing.

Returns a linux_dmabuf_attributes object, owned by the backend;
or NULL if the buffer doesn't represent a linux-dmabuf buffer.